+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 -   Fixed an error when trying to load the plugin settings page. ([#3](https://github.com/michaelhue/craft-breakpoint/issues/3))
+-   Fixed compatibility with Craft `3.4.28` and newer.
 
 ## 2.0.0 - 2019-07-11